Change architecture in comment package
Remove relations to other modules
