--- src/main/java/demo/codeexample/comment/CommentEntity.java
@@ -1,64 +1,0 @@
-package demo.codeexample.comment;
-
-import demo.codeexample.task.TaskEntity;
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.hibernate.proxy.HibernateProxy;
-
-import java.time.LocalDateTime;
-import java.util.Objects;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@Entity
-public class CommentEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Setter(AccessLevel.NONE)
-    private Long id;
-
-    @Column(nullable = false, length = 300)
-    private String content;
-
-    @Column(nullable = false)
-    private Long writerId;
-
-    @ManyToOne(optional = false, fetch = FetchType.LAZY)
-    @JoinColumn(name = "task_id")
-    private TaskEntity task;
-
-    @Column(nullable = false, updatable = false)
-    private LocalDateTime createdAt;
-
-    public CommentEntity(String content, Long writerId, TaskEntity task) {
-        this.content = content;
-        this.writerId = writerId;
-        this.task = task;
-    }
-
-    @PrePersist
-    protected void onCreate() {
-        this.createdAt = LocalDateTime.now();
-    }
-
-    @Override
-    public final boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null) return false;
-        Class<?> oEffectiveClass = o instanceof HibernateProxy ? ((HibernateProxy) o).getHibernateLazyInitializer().getPersistentClass() : o.getClass();
-        Class<?> thisEffectiveClass = this instanceof HibernateProxy ? ((HibernateProxy) this).getHibernateLazyInitializer().getPersistentClass() : this.getClass();
-        if (thisEffectiveClass != oEffectiveClass) return false;
-        CommentEntity that = (CommentEntity) o;
-        return getId() != null && Objects.equals(getId(), that.getId());
-    }
-
-    @Override
-    public final int hashCode() {
-        return this instanceof HibernateProxy ? ((HibernateProxy) this).getHibernateLazyInitializer().getPersistentClass().hashCode() : getClass().hashCode();
-    }
-}
