+++ src/main/java/demo/codeexample/comment/CommentFacade.java
@@ -1,0 +1,5 @@
+package demo.codeexample.comment;
+
+public interface CommentFacade {
+
+}
