+++ src/main/java/demo/codeexample/comment/application/CommentService.java
@@ -1,0 +1,33 @@
+package demo.codeexample.comment.application;
+import demo.codeexample.comment.CommentDto;
+import demo.codeexample.comment.domain.CommentRepository;
+import demo.codeexample.comment.CreateCommentDto;
+import demo.codeexample.comment.domain.CommentEntity;
+import demo.codeexample.task.TaskEntity;
+import demo.codeexample.task.TaskRepository;
+import lombok.AllArgsConstructor;
+import org.modelmapper.ModelMapper;
+import org.springframework.stereotype.Service;
+
+@Service
+@AllArgsConstructor
+public class CommentService {
+
+    private final CommentRepository commentRepository;
+    private final TaskRepository taskRepository;
+    private final ModelMapper modelMapper;
+    //private final "Some kind of Authorization service" with saved user
+
+    //Add
+    public CommentDto createComment(CreateCommentDto createCommentDto) {
+
+        // Long creatorId = Any AuthService
+
+        CommentEntity commentEntity = modelMapper.map(createCommentDto, CommentEntity.class);
+        // commentEntity.setCreatorId(creatorId)
+        commentRepository.save(commentEntity);
+
+        return modelMapper.map(commentEntity, CommentDto.class);
+    }
+
+}
